added tfsenc_config.py
