+++ .gitignore
@@ -7,6 +7,7 @@ __pycache__
 data/*
 logs
 results/*
+misc/*
 
 !.gitkeep
 .vscode
